+++ apps/app/src/features/rate-limiter/middleware/rate-limiter-mongo-client.ts
@@ -1,0 +1,11 @@
+import { connection } from 'mongoose';
+import { type IRateLimiterMongoOptions, RateLimiterMongo } from 'rate-limiter-flexible';
+
+import { DEFAULT_DURATION_SEC } from '../config';
+
+const opts: IRateLimiterMongoOptions = {
+  storeClient: connection,
+  duration: DEFAULT_DURATION_SEC, // set default value
+};
+
+export const rateLimiter = new RateLimiterMongo(opts);
